Prod deployment dev (#6)
* Update: Update validator URL construction

Validator function was previously pulling the URL from the request. URL is now created from the headers and a environment variable to pull missing part of url that can't be pulled from the header.

* Add: New main entry point

Added a new main entry point file for the cloud functions to start the uvicorn app

---------

Co-authored-by: redjordan <jordan@delpilar.net>