Docker Deoplyment Dev (#7)
* Update: Updated .gitignore

* Add: Created Dockerfile

Created new Dockerfile to allow for building of docker image. This will allow for using docker for both testing and deployment in the future.

* Add: docker-compose.yaml

Create example docker-compose.yaml. This docker-compose will run and deploy with no issues in its default setup but lets user's set the container up as they wish.

* Add: example env file

An example .env file acting as a base for the .env file that users will need to deploy the chiron relay via docker or to host it locally if desired. This contains no real secrets.

* Update: Updated README.md

Update the readme to include steps needed to deploy both with docker and locally without docker.
Also minor spelling, grammar and formatting fixes throughout.

---------

Co-authored-by: redjordan <jordan@delpilar.net>